Give Interpolation.Easing its own file
